requirements.txt
apscheduler
numpy
pandas 
matplotlib
seaborn
zhplot
scikit-learn
onnx
onnxruntime
redis
taospy
bayesian-optimization
fastapi
uvicorn
pyyaml
psutil
